Refactor user.action component
