+++ components/user-mgt/org.wso2.carbon.identity.user.action/src/main/java/org/wso2/carbon/identity/user/action/api/service/UserActionExecutor.java
@@ -16,12 +16,12 @@
  * under the License.
  */
 
-package org.wso2.carbon.identity.user.action.service;
+package org.wso2.carbon.identity.user.action.api.service;
 
 import org.wso2.carbon.identity.action.execution.exception.ActionExecutionException;
 import org.wso2.carbon.identity.action.execution.model.ActionExecutionStatus;
 import org.wso2.carbon.identity.action.execution.model.ActionType;
-import org.wso2.carbon.identity.user.action.service.model.UserActionContext;
+import org.wso2.carbon.identity.user.action.api.model.UserActionContext;
 
 /**
  * Interface for User Action Executor.
