+++ components/user-mgt/org.wso2.carbon.identity.user.action/src/main/java/org/wso2/carbon/identity/user/action/internal/factory/UserActionExecutorFactory.java
@@ -16,10 +16,10 @@
  * under the License.
  */
 
-package org.wso2.carbon.identity.user.action.core.factory;
+package org.wso2.carbon.identity.user.action.internal.factory;
 
 import org.wso2.carbon.identity.action.execution.model.ActionType;
-import org.wso2.carbon.identity.user.action.service.UserActionExecutor;
+import org.wso2.carbon.identity.user.action.api.service.UserActionExecutor;
 
 import java.util.EnumMap;
 import java.util.Map;
